Add endpoints like "findAllMediaType"; required some used packages
